move saves and coordinates out of tabuleiro.c
