+++ src/jogo/comandos.c
@@ -6,6 +6,9 @@
 
 #include "helpers/strs.h"
 #include "helpers/history.h"
+#include "helpers/coords.h"
+
+#include "jogo/saves.h"
 #include "jogo/comandos.h"
 #include "jogo/tabuleiro.h"
 #include "types.h"
